requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
captcha==0.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
frozenlist==1.4.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
gspread==6.0.2
gspread_asyncio==2.0.0
idna==3.7
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
oauthlib==3.2.2
orjson==3.10.3
pillow==10.3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
python-dotenv==1.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==2.2.0
SQLAlchemy==2.0.30
StrEnum==0.4.15
typing_extensions==4.11.0
urllib3==2.2.1
uvloop==0.19.0
yarl==1.9.4
